Update:
- Update to changes in rest-api, return long instead of int for counting.
- Fix concrete SearchResult classes
- Update craetion of DatasetOutput
- Update Spring dependencies

+++ dao/src/main/java/org/n52/series/db/dao/AbstractDao.java
@@ -126,18 +126,18 @@ public abstract class AbstractDao<T> implements GenericDao<T, Long> {
     }
 
     @Override
-    public Integer getCount(DbQuery query) throws DataAccessException {
+    public Long getCount(DbQuery query) throws DataAccessException {
         if (!DataModelUtil.isEntitySupported(getEntityClass(), session)) {
-            return 0;
+            return 0L;
         }
         Criteria criteria = getDefaultCriteria(query).setProjection(Projections.rowCount());
         Object result = criteria.uniqueResult();
         if (result == null) {
             String sql = DataModelUtil.getSqlString(criteria);
             LOGGER.error("Please review query: {}", sql);
-            return 0;
+            return 0L;
         }
-        return ((Long) result).intValue();
+        return (Long) result;
     }
 
     protected <I extends I18nEntity> Criteria i18n(Class<I> clazz, Criteria criteria, DbQuery query) {
